+++ jkube-kit/jkube-kit-spring-boot/src/main/java/org/eclipse/jkube/springboot/enricher/SpringBootHealthCheckEnricher.java
@@ -20,6 +20,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.eclipse.jkube.kit.common.Configs;
 import org.eclipse.jkube.kit.common.util.JKubeProjectUtil;
+import org.eclipse.jkube.kit.common.util.ProjectClassLoaders;
 import org.eclipse.jkube.kit.common.util.SpringBootConfiguration;
 import org.eclipse.jkube.kit.common.util.SpringBootUtil;
 import org.eclipse.jkube.kit.enricher.api.JKubeEnricherContext;
@@ -38,9 +39,14 @@ public class SpringBootHealthCheckEnricher extends AbstractHealthCheckEnricher {
 
     public static final String ENRICHER_NAME = "jkube-healthcheck-spring-boot";
 
-    protected static final String[] REQUIRED_CLASSES = {
-            "org.springframework.boot.actuate.health.HealthIndicator",
-            "org.springframework.web.context.support.GenericWebApplicationContext"
+    protected static final String[] REQUIRED_CLASSES_SPRING_BOOT = {
+        "org.springframework.boot.actuate.health.HealthIndicator",
+        "org.springframework.web.context.support.GenericWebApplicationContext"
+    };
+
+    protected static final String[] REQUIRED_CLASSES_SPRING_BOOT_4 = {
+        "org.springframework.boot.health.contributor.HealthIndicator",
+        "org.springframework.web.context.support.GenericWebApplicationContext"
     };
 
     private static final String SCHEME_HTTPS = "HTTPS";
@@ -97,8 +103,10 @@ public class SpringBootHealthCheckEnricher extends AbstractHealthCheckEnricher {
 
     protected Probe discoverSpringBootHealthCheck(Integer initialDelay, Integer period, Integer timeout, Integer failureTh, Integer successTh, String suffix) {
         try {
-            if (getContext().getProjectClassLoaders().isClassInCompileClasspath(true, REQUIRED_CLASSES)) {
-                return buildProbe(initialDelay, period, timeout, failureTh, successTh, suffix);
+            ProjectClassLoaders projectClassLoaders = getContext().getProjectClassLoaders();
+            if (projectClassLoaders.isClassInCompileClasspath(true, REQUIRED_CLASSES_SPRING_BOOT) ||
+                projectClassLoaders.isClassInCompileClasspath(true, REQUIRED_CLASSES_SPRING_BOOT_4)) {
+                    return buildProbe(initialDelay, period, timeout, failureTh, successTh, suffix);
             }
         } catch (Exception ex) {
             log.error("Error while reading the spring-boot configuration", ex);
@@ -119,11 +127,11 @@ public class SpringBootHealthCheckEnricher extends AbstractHealthCheckEnricher {
         String scheme;
         String prefix;
         if (usingManagementPort) {
-            scheme = StringUtils.isNotBlank(springBootConfiguration.getManagementKeystore()) ? SCHEME_HTTPS : SCHEME_HTTP;
+            scheme = StringUtils.isNotBlank(springBootConfiguration.getManagementKeystore()) && springBootConfiguration.isManagementSslEnabled() ? SCHEME_HTTPS : SCHEME_HTTP;
             prefix = StringUtils.isNotBlank(springBootConfiguration.getManagementContextPath()) ?
               springBootConfiguration.getManagementContextPath() : "";
         } else {
-            scheme = StringUtils.isNotBlank(springBootConfiguration.getServerKeystore()) ? SCHEME_HTTPS : SCHEME_HTTP;
+            scheme = StringUtils.isNotBlank(springBootConfiguration.getServerKeystore()) && springBootConfiguration.isServerSslEnabled() ? SCHEME_HTTPS : SCHEME_HTTP;
             if (hasSpringWebFluxDependency(getContext().getProject()) && StringUtils.isNotBlank(springBootConfiguration.getWebFluxBasePath())) {
                 prefix = springBootConfiguration.getWebFluxBasePath();
             } else if (StringUtils.isNotBlank(springBootConfiguration.getServerContextPath())) {