+++ CHANGELOG.md
@@ -26,6 +26,9 @@ Usage:
 * Fix #3781: Native Generator does not work in Windows system, always finds multiple native executable
 * Fix #2286: Remove Guava dependency where ever possible
 * Fix #3732: ECR registry Auth with AWS SDK java v2
+* Fix #3809: Actuator liveness and readiness probe not getting generated with Spring boot 4.x.x
+* Fix #3707: Setting readOnly flag in VolumeConfig has no effect
+* Fix #1458: Consideration of "ssl.enabled" properties to enable liveness/readiness probe for Spring Boot Actuator
 
 ### 1.18.2 (2025-11-03)
 * Fix #3750: Remove unneeded XMLUtil.createNewDocument method